pos-service/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
Deprecated==1.2.15
googleapis-common-protos==1.66.0
grpcio==1.68.0
idna==3.10
importlib_metadata==8.5.0
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-requests==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
packaging==24.2
protobuf==5.28.3
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.17.0
zipp==3.21.0
